refactor(main-cli): move helper function into the module
